Fix flaky test IndexShardTests.testLocalDirectoryContains
This test is breaking for WindowsFS only. Moving it to a separate file where it is skipped on WindowsFS.
